Handle corner case in intersection computation
We have two functions which compute the intersection of a line with a
circular/rectangular shape. If needed, the line segment is extended to
ensure an intersection exists. However, if both points defining the line
are the same, the computation gives NaN as the result, which turns into
zero as an integer.

Thus, when you create two graphical elements, connect them with a line
and then move them onto each other, the line between them appears in the
top-left corner of the canvas.

Fix this by explicitly handling this special case in intersection
computation. We define the intersection in this case to be in the center
of the shape. This means that in the above case, the connecting line is
now drawn as an arrow-head in the middle of the shape.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>