Make it possible to abort module executions
When a module is run the "Run" button now turns into "Abort". Pressing
this button will send an interruption to the thread executing the
module. How well this works depends on the module, but at least for
InterruptibleModule instances this should work.

Since this is now about interactions between multiple threads, we have a
synchronisation problem. The difficulty here is that when we abort the
module before it actually started running (e.g. because the queue of our
executor filled up), the Runnable for the module will not be invoked.
However, if the Runnable was already invoked, we want to wait for the
module to really be done.

This complexity is encapsulated in an inner class whose job it is to
ensure that some cleanup is done when the module really stopped
execution. The cleanup here is mainly turning the "Abort" button back
into "Run" and to clean up the internal state so that a new module
invocation can be done.

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>