Improve handling of optional parameters
Obviously, leaving an input box empty means that the optional parameter
is not provided. This also works as intended. However, if you enter
something and then erase it again, the empty string is used as the
parameter. Since none of our parameter transformations do anything
useful for the empty string, this then fails.

This commit fixes this by interpreting the empty string as "the
parameter was not given".

Signed-off-by: Uli Schlachter <uli.schlachter@informatik.uni-oldenburg.de>